Move BatchJob model to the aws app.
