+++ project/vision_backend/admin.py
@@ -1,11 +1,5 @@
 from django.contrib import admin
-from .models import Classifier, BatchJob
-
-
-@admin.register(BatchJob)
-class BatchJobAdmin(admin.ModelAdmin):
-    list_display = (
-        'create_date', 'status', 'batch_token', 'internal_job', 'spec_level')
+from .models import Classifier
 
 
 @admin.register(Classifier)
